Login errors READ DESCRIPTION
If a user tries to login and is incorrectly logged in when we pass them back to the login screen I added an error message that will be posted at the very top of the page.  It's kind of like what twitter does when you don't authenticate.  I think you should just style this up at the top and we could make it slide down or something to keep it simple yet cool looking.